requirements.txt
Django==1.9.5
flake8==2.5.4
Pillow==3.2.0
wheel==0.24.0
django-ckeditor==5.0.3
djangorestframework==3.3.3
pytz
